# performance optimization of primelib.isPrime(number)