only loop over the odd numbers.


## performance test:

### old version:

In [1]: from primelib import isPrime

In [2]: %timeit isPrime(99999999999999997)
50.9 s ± 31.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)


### new version:

In [1]: from primelib import isPrime

In [2]: %timeit isPrime(99999999999999997)
32.4 s ± 149 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)